+++ src/test/java/gift/repository/JpaCategoryTest.java
@@ -1,8 +1,6 @@
-package gift;
+package gift.repository;
 
 import gift.model.Category;
-import gift.model.Product;
-import gift.repository.CategoryRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
