+++ src/test/java/gift/repository/JpaMemberTest.java
@@ -1,7 +1,6 @@
-package gift;
+package gift.repository;
 
 import gift.model.Member;
-import gift.repository.MemberRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
