+++ src/test/java/gift/repository/JpaProductTest.java
@@ -1,9 +1,7 @@
-package gift;
+package gift.repository;
 
 import gift.model.Category;
 import gift.model.Product;
-import gift.repository.CategoryRepository;
-import gift.repository.ProductRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
