+++ src/test/java/gift/repository/JpaWishTest.java
@@ -1,11 +1,8 @@
-package gift;
+package gift.repository;
 
 import gift.model.Member;
 import gift.model.Product;
 import gift.model.Wish;
-import gift.repository.MemberRepository;
-import gift.repository.ProductRepository;
-import gift.repository.WishRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
